Refactor video encoding service:
- Migrate video-related logic into a modular structure under `src`.
- Extract reusable utilities and error handling into dedicated modules: `exceptions`, `services`, and `schemas`.
- Modularize `rabbit_client` logic into `main.py` and update `Dockerfile` to reflect the new entrypoint.
- Replace relative imports with absolute imports under `src` for consistency.
- Standardize error handling around custom exceptions for robustness and clarity.
- Introduce consolidated `VideoProperties` schema for improved type definition and validation.
